requirements.txt
# OCR Service Dependencies

# Core OCR Engine
rapidocr-onnxruntime==1.4.4

# ONNX Runtime (CPU execution)
onnxruntime==1.20.1

# Web Framework
fastapi==0.115.7
uvicorn[standard]==0.40.0

# Production WSGI Server
gunicorn==21.2.0

# Image Processing
Pillow==11.0.0
opencv-python==4.9.0.80
numpy==1.26.4

# Data Processing
pydantic==2.9.2
PyYAML==6.0.1

# Geometry Processing (OCR post-processing)
pyclipper==1.3.0.post5
shapely==2.0.5

# PDF Processing
PyMuPDF==1.24.12

# FastAPI File Upload Support
python-multipart==0.0.12

# Testing
requests==2.32.3
